Use `recognizes?` and `prepare` pattern
