+++ apps/remote_control/lib/lexical/remote_control/code_mod/rename/prepare.ex
@@ -1,12 +1,10 @@
 defmodule Lexical.RemoteControl.CodeMod.Rename.Prepare do
-  alias Lexical.Ast
   alias Lexical.Ast.Analysis
   alias Lexical.Document.Position
   alias Lexical.Document.Range
   alias Lexical.Formats
   alias Lexical.RemoteControl.CodeIntelligence.Entity
   alias Lexical.RemoteControl.CodeMod.Rename
-  alias Lexical.RemoteControl.Search.Store
 
   require Logger
 
@@ -28,55 +26,22 @@ defmodule Lexical.RemoteControl.CodeMod.Rename.Prepare do
     end
   end
 
+  @renaming_modules [Rename.Module]
   @spec resolve(Analysis.t(), Position.t()) ::
-          {:ok, {atom(), atom()} | {atom(), tuple()}, Range.t()} | {:error, term()}
+          {:ok, {atom(), atom()}, Range.t()} | {:error, tuple() | atom()}
   def resolve(%Analysis{} = analysis, %Position{} = position) do
-    case do_resolve(analysis, position) do
-      {:ok, {:module, _module}, _range} ->
-        {module, range} = surround_the_whole_module(analysis, position)
-
-        if cursor_at_declaration?(module, range) do
-          {:ok, {:module, module}, range}
-        else
-          {:error, {:unsupported_location, :module}}
+    prepare_result =
+      Enum.find_value(@renaming_modules, fn module ->
+        if module.recognizes?(analysis, position) do
+          module.prepare(analysis, position)
         end
+      end)
 
-      other ->
-        other
-    end
+    prepare_result || handle_unsupported_entity(analysis, position)
   end
 
-  defp surround_the_whole_module(analysis, position) do
-    # When renaming occurs, we want users to be able to choose any place in the defining module,
-    # not just the last local module, like: `defmodule |Foo.Bar do` also works.
-    {:ok, %{end: {_end_line, end_character}}} = Ast.surround_context(analysis, position)
-    end_position = %{position | character: end_character - 1}
-    {:ok, {:module, module}, range} = do_resolve(analysis, end_position)
-    {module, range}
-  end
-
-  defp cursor_at_declaration?(module, rename_range) do
-    case Store.exact(module, type: :module, subtype: :definition) do
-      {:ok, [definition]} ->
-        rename_range == definition.range
-
-      _ ->
-        false
-    end
-  end
-
-  @renamable_modules [Rename.Module]
-
-  defp do_resolve(%Analysis{} = analysis, %Position{} = position) do
-    apply_resolve = fn module ->
-      case module.resolve(analysis, position) do
-        {:ok, _, _} = result -> result
-        _ -> false
-      end
-    end
-
-    with :error <- Enum.find_value(@renamable_modules, :error, apply_resolve),
-         {:ok, other, _} <- Entity.resolve(analysis, position) do
+  defp handle_unsupported_entity(analysis, position) do
+    with {:ok, other, _range} <- Entity.resolve(analysis, position) do
       Logger.info("Unsupported entity for renaming: #{inspect(other)}")
       {:error, {:unsupported_entity, elem(other, 0)}}
     end
